prepared constructors for database entities

+++ app/src/main/java/com/example/water_your_plants_app/database/tables/Table_Task.java
@@ -7,5 +7,5 @@ import androidx.room.PrimaryKey;
 public class Table_Task {
 
     @PrimaryKey(autoGenerate = true)
-    public int task_id;
+    public long task_id;
 }
